chore: cleanup file naming and organize imports
