+++ src/features/consultation/model/consultation.schemas.ts
